chore: tmpChat, tmpMemo Id type 변경

+++ src/main/java/corecord/dev/domain/user/entity/User.java
@@ -29,9 +29,9 @@ public class User extends BaseEntity {
     private Status status;
 
     @Column
-    private Integer tmpChat;
+    private Long tmpChat;
 
     @Column
-    private Integer tmpMemo;
+    private Long tmpMemo;
 
 }
